refactors to components using module pattern. splits all functions
